requirements.txt
sounddevice
numpy
langchain
langchain-community
python-dotenv
simpleaudio
